api-srcei/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pymysql
mysql-connector-python